backend/requirements.txt
fastapi==0.115.0
numpy==1.26.4
pandas==2.2.2
pydantic==2.9.2
pymongo==3.10.0
python-dotenv==1.0.1
uvicorn==0.31.0
scikit-learn==1.5.0
bcrypt==4.2.0
keras==3.3.3
dnspython==1.16.0
joblib==1.4.2
tensorflow==2.16.1
git+https://github.com/rongardF/tvdatafeed.git